# docs/api/dsg-openapi.yaml
openapi: 3.0.3
info:
  title: DSG Admin API
  description: |
    Directory Sync Gateway configuration API (wiki section 2.6).
    Service Web (SCL) → JWL → DSG. Phase 1 planning contract.
  version: 0.1.0
  contact:
    name: DSG Engineering
servers:
  - url: http://localhost:8080/dsg/v1
    description: Local API
  - url: https://{host}/dsg/v1
    description: Deployed
    variables:
      host:
        default: api.example.com
...
paths:
...
  /{accountId}/jobs:
    parameters:
      - $ref: '#/components/parameters/accountId'
    get:
      tags: [jobs]
      summary: List sync job history for account (Service Web)
      operationId: listJobs
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        '200':
          description: Job history
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobHistoryResponse'
    post:
      tags: [jobs]
      summary: Trigger sync job (full, incremental, on-demand)
      operationId: createJob
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateJobRequest'
      responses:
        '202':
          description: Job accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobResponse'
        '409':
          description: Another job already running for this account
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /{accountId}/jobs/latest/report:
    parameters:
      - $ref: '#/components/parameters/accountId'
    get:
      tags: [jobs]
      summary: Get report for the most recent job (sync status cards)
      operationId: getLatestJobReport
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobReportResponse'
        '404':
          description: No jobs for account
  /{accountId}/jobs/{jobId}/report:
    parameters:
      - $ref: '#/components/parameters/accountId'
      - name: jobId
        in: path
        required: true
        schema:
          type: string
    get:
      tags: [jobs]
      summary: Get sync job report (P0-9 Phase 1)
      operationId: getJobReport
      responses:
        '200':
          description: Report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobReportResponse'
        '404':
          description: Job not found for account
components:
  parameters:
    accountId:
      name: accountId
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
...
    CreateJobRequest:
      type: object
      required: [jobType]
      properties:
        jobType:
          type: string
          enum: [FULL, INCREMENTAL, ON_DEMAND]
        externalUserIds:
          type: array
          items:
            type: string
          description: On-demand subset or failed-user retry
    JobResponse:
      type: object
      properties:
        jobId:
          type: string
        state:
          type: string
    JobReportResponse:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          type: string
          enum: [FULL, INCREMENTAL, ON_DEMAND]
        syncDirection:
          type: string
          enum: [DIR_TO_RC, RC_TO_DIR]
        state:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        successCount:
          type: integer
        failedCount:
          type: integer
        failures:
          type: array
          items:
            $ref: '#/components/schemas/JobFailure'
    JobFailure:
      type: object
      properties:
        externalId:
          type: string
        operation:
          type: string
        comment:
          type: string
    JobHistoryResponse:
      type: object
      properties:
        jobs:
          type: array
          items:
            $ref: '#/components/schemas/JobSummaryResponse'
    JobSummaryResponse:
      type: object
      properties:
        jobId:
          type: string
        jobType:
          type: string
        syncDirection:
          type: string
        state:
          type: string
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        successCount:
          type: integer
        failedCount:
          type: integer